Always use the ET_DYN codepath, avoiding shifting load start
The executable codepath of operations like --set-rpath can cause the
first address of a binary to move earlier, causing segfaults on startup
of some binaries on e.g.  riscv64 where the default base address matches
the kernel's mmap_min_addr.

PIE executables, which are the great majority, already use the library
path, so this must be pretty safe for the remaining few.